+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - ASB+CONTROLLER: New `set-external-bitcoin-redeem-address` command with parameter `address` (string) allows you to udpate the external bitcoin redeem address at runtime.
    The change is persisted to the config file. Go back to using the internal Bitcoin wallet by using `clear-external-bitcoin-redeem-address`.
 
+## [4.6.0] - 2026-05-13
+
+- GUI: Exolix has become a community supporter. They support development and provide a reliable market maker. They are visually highlighted in the UI.
+
 ## [4.5.5] - 2026-05-13
 
 ## [4.5.4] - 2026-05-12
@@ -942,7 +946,8 @@ It is possible to migrate critical data from the old db to the sqlite but there 
 - Fixed an issue where Alice would not verify if Bob's Bitcoin lock transaction is semantically correct, i.e. pays the agreed upon amount to an output owned by both of them.
   Fixing this required a **breaking change** on the network layer and hence old versions are not compatible with this version.
 
-[unreleased]: https://github.com/eigenwallet/core/compare/4.5.5...HEAD
+[unreleased]: https://github.com/eigenwallet/core/compare/4.6.0...HEAD
+[4.6.0]: https://github.com/eigenwallet/core/compare/4.5.5...4.6.0
 [4.5.5]: https://github.com/eigenwallet/core/compare/4.5.4...4.5.5
 [4.5.4]: https://github.com/eigenwallet/core/compare/4.5.3...4.5.4
 [4.5.3]: https://github.com/eigenwallet/core/compare/4.5.2...4.5.3